requirements.txt
# Core Flask Framework
flask==2.3.3
flask-socketio==5.3.6
flask-cors==4.0.0
flask-session==0.5.0
flask-sqlalchemy==3.0.5

# Core FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Socket.IO for WebSockets (with eventlet for async support)
python-socketio==5.10.0
eventlet==0.35.2
dnspython==2.6.1  # Required by eventlet
pyopenssl==24.1.0  # Compatible with eventlet 0.35.2
cryptography==42.0.5  # Compatible with pyopenssl

# Database & ORM
psycopg[binary,pool]==3.1.19
psycopg2-binary==2.9.7
psycopg-pool==3.2.1
sqlalchemy==2.0.23

# HTTP & API
requests==2.31.0
httpx==0.25.2

# Authentication & Security
pyjwt==2.8.0
python-multipart==0.0.6

# Environment & Configuration
python-dotenv==1.0.0

# PDF Generation
reportlab==4.0.4

# Excel Processing
openpyxl==3.1.2

# Data Processing
pandas==2.0.3
numpy==1.24.3

# Utilities
python-dateutil==2.8.2
pytz==2023.3
psutil==5.9.8

# Development & Testing (optional)
pytest==7.4.0
pytest-asyncio==0.21.1
